M2 point 0: widen fontdesc_StringBoundingBox .ch signature to long* to match StringSize/TextSize siblings; fix figotext.c's int locals to long to match
FossilOrigin-Name: 0a6cf595ef5142f5c35b77927f7e93122d91c10dc6cf15dcc02a6b2e08d173a5

+++ src/atk/basics/common/fontdesc.c
@@ -361,8 +361,8 @@ long fontdesc__StringBoundingBox(font, graphic, string, width, height)
 struct fontdesc *font;
 struct graphic *graphic;
 char *string;
-int *width;
-int *height;
+long *width;
+long *height;
 {
   int w, a, d, ascent, descent, junk;
   register short *fwt, *fht;
